refactor : Exception 부분 네이밍 S3 -> Image 로 변경
