+++ backend/src/main/java/com/mapbefine/mapbefine/image/exception/ImageErrorCode.java
@@ -3,7 +3,7 @@ package com.mapbefine.mapbefine.image.exception;
 import lombok.Getter;
 
 @Getter
-public enum S3ErrorCode {
+public enum ImageErrorCode {
 
     ILLEGAL_IMAGE_FILE_EXTENSION("10000", "이미지 파일이 아닙니다."),
     IMAGE_FILE_IS_NULL("10001", "이미지가 선택되지 않았습니다.")
@@ -12,7 +12,7 @@ public enum S3ErrorCode {
     private final String code;
     private final String message;
 
-    S3ErrorCode(String code, String message) {
+    ImageErrorCode(String code, String message) {
         this.code = code;
         this.message = message;
     }
