+++ backend/src/main/java/com/mapbefine/mapbefine/image/exception/ImageException.java
@@ -3,11 +3,11 @@ package com.mapbefine.mapbefine.image.exception;
 import com.mapbefine.mapbefine.common.exception.BadRequestException;
 import com.mapbefine.mapbefine.common.exception.ErrorCode;
 
-public class S3Exception {
+public class ImageException {
 
-    public static class S3BadRequestException extends BadRequestException {
+    public static class ImageBadRequestException extends BadRequestException {
 
-        public S3BadRequestException(S3ErrorCode errorCode) {
+        public ImageBadRequestException(ImageErrorCode errorCode) {
             super(new ErrorCode<>(errorCode.getCode(), errorCode.getMessage()));
         }
