+++ backend/src/main/java/com/mapbefine/mapbefine/pin/application/PinCommandService.java
@@ -3,7 +3,7 @@ package com.mapbefine.mapbefine.pin.application;
 import static com.mapbefine.mapbefine.pin.exception.PinErrorCode.FORBIDDEN_PIN_CREATE_OR_UPDATE;
 import static com.mapbefine.mapbefine.pin.exception.PinErrorCode.ILLEGAL_PIN_ID;
 import static com.mapbefine.mapbefine.pin.exception.PinErrorCode.ILLEGAL_PIN_IMAGE_ID;
-import static com.mapbefine.mapbefine.image.exception.S3ErrorCode.IMAGE_FILE_IS_NULL;
+import static com.mapbefine.mapbefine.image.exception.ImageErrorCode.IMAGE_FILE_IS_NULL;
 import static com.mapbefine.mapbefine.topic.exception.TopicErrorCode.ILLEGAL_TOPIC_ID;
 
 import com.mapbefine.mapbefine.auth.domain.AuthMember;
@@ -23,7 +23,7 @@ import com.mapbefine.mapbefine.pin.dto.request.PinUpdateRequest;
 import com.mapbefine.mapbefine.pin.exception.PinException.PinBadRequestException;
 import com.mapbefine.mapbefine.pin.exception.PinException.PinForbiddenException;
 import com.mapbefine.mapbefine.image.application.ImageService;
-import com.mapbefine.mapbefine.image.exception.S3Exception.S3BadRequestException;
+import com.mapbefine.mapbefine.image.exception.ImageException.ImageBadRequestException;
 import com.mapbefine.mapbefine.topic.domain.Topic;
 import com.mapbefine.mapbefine.topic.domain.TopicRepository;
 import com.mapbefine.mapbefine.topic.exception.TopicException.TopicBadRequestException;
@@ -165,7 +165,7 @@ public class PinCommandService {
 
     private void addImageToPin(MultipartFile image, Pin pin) {
         if (Objects.isNull(image)) {
-            throw new S3BadRequestException(IMAGE_FILE_IS_NULL);
+            throw new ImageBadRequestException(IMAGE_FILE_IS_NULL);
         }
 
         String imageUrl = imageService.upload(image);
