+++ backend/src/test/java/com/mapbefine/mapbefine/pin/application/PinCommandServiceTest.java
@@ -25,7 +25,7 @@ import com.mapbefine.mapbefine.pin.dto.request.PinUpdateRequest;
 import com.mapbefine.mapbefine.pin.dto.response.PinDetailResponse;
 import com.mapbefine.mapbefine.pin.dto.response.PinImageResponse;
 import com.mapbefine.mapbefine.pin.exception.PinException.PinForbiddenException;
-import com.mapbefine.mapbefine.image.exception.S3Exception.S3BadRequestException;
+import com.mapbefine.mapbefine.image.exception.ImageException.ImageBadRequestException;
 import com.mapbefine.mapbefine.topic.TopicFixture;
 import com.mapbefine.mapbefine.topic.domain.Topic;
 import com.mapbefine.mapbefine.topic.domain.TopicRepository;
@@ -245,7 +245,7 @@ class PinCommandServiceTest {
 
         // when then
         assertThatThrownBy(() -> pinCommandService.addImage(authMember, imageNullRequest))
-                .isInstanceOf(S3BadRequestException.class);
+                .isInstanceOf(ImageBadRequestException.class);
     }
 
     @Test
